Move NumeralInput components into elements folder as they are not used elsewhere
