+++ app/packs/src/components/elements/reactions/schemeTab/Material.js
@@ -4,7 +4,7 @@ import { Radio, FormControl, Button, InputGroup, OverlayTrigger, Tooltip } from 
 import { DragSource, DropTarget } from 'react-dnd';
 import { compose } from 'redux';
 import DragDropItemTypes from 'src/components/DragDropItemTypes';
-import NumeralInputWithUnitsCompo from 'src/components/NumeralInputWithUnitsCompo';
+import NumeralInputWithUnitsCompo from 'src/components/elements/NumeralInputWithUnitsCompo';
 import SampleName from 'src/components/common/SampleName';
 import ElementActions from 'src/stores/alt/actions/ElementActions';
 import { UrlSilentNavigation, SampleCode } from 'src/utilities/ElementUtils';
